# specs/007-pagos-seguridad/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Rasero — Pagos y Seguridad
  version: 0.1.0
  description: |
    Contrato del módulo 007-pagos-seguridad. **Cuatro preocupaciones acotadas del encargo**
    (FR-033), cada una con su recurso:

    - `/pagos/medios` y `/pagos/cobertura` — catálogo de medios de pago y su aceptación HISTÓRICA
      por sucursal, más la cuota de intención de compra no atendida cuando un cliente no pudo pagar
      como quería (FR-001 a FR-007). Es la métrica que la **Lectura Crítica n.º 4** nombra; NO es
      `consulta_no_atendida` de 001 ("no hay producto") ni faltante de inventario.
    - `/pagos/terminales` — registro de datáfonos físicos, su versión de firmware y los indicadores
      DERIVADOS "desactualizada" / "expuesta a clonación" / "versión de referencia desconocida"
      (FR-008 a FR-015). El sistema NUNCA deshabilita una terminal por una señal (FR-012).
    - `/pagos/tokens` y `/pagos/ventas/{id_venta}/pago` — tokenización: cuando una `venta` de 001 se
      cobra con tarjeta, se conserva ÚNICAMENTE {token opaco, últimos 4 dígitos, marca, tipo,
      terminal, referencia de venta}; el PAN, el CVV y los datos de banda/chip NUNCA se almacenan,
      registran ni transmiten (FR-016 a FR-022). IDEMPOTENTE por `id_venta` (un token por cobro).
    - `/pagos/bitacora` — rastro de SOLO ANEXADO de los hechos de pago (FR-023 a FR-028, Principio
      IV).

    Reglas transversales que aplican a todo el contrato:

...
      número (FR-018, SC-008).
...
      FR-034).
...

...
servers:
...
paths:
...
  /pagos/terminales:
    post:
      tags: [terminales]
      summary: Registrar una terminal de pago física
      description: FR-008, FR-015. Reservado al encargado. Anexa `terminal_registrada` a la bitácora.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identificador, modelo, id_sucursal, version_firmware, id_operador]
              properties:
                identificador: { type: string }
                modelo: { type: string }
                id_sucursal: { type: integer }
                version_firmware:
                  type: string
                  pattern: '^[0-9]+\.[0-9]+\.[0-9]+$'
                fecha_ultima_actualizacion_firmware: { type: string, format: date, nullable: true }
                id_operador: { type: integer, description: 'Debe ser `es_encargado` (FR-015).' }
      responses:
        '201':
          description: Terminal registrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TerminalPago' }
        '400':
          description: '`pagos_version_firmware_invalida`, `pagos_identificador_duplicado`, `pagos_operador_no_encargado`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: '`pagos_sucursal_no_existe`, `pagos_operador_no_existe`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [terminales]
      summary: Estado de las terminales de una sucursal, con los indicadores de firmware derivados
      description: |
        FR-010, FR-011, FR-014. Los indicadores `desactualizada`, `expuesta_a_clonacion` y
        `version_referencia_desconocida` se CALCULAN al leer contra `ULTIMA_VERSION_FIRMWARE` y
        `LISTA_FIRMWARE_VULNERABLE` (config). No hay tabla de indicadores. Ninguna señal deshabilita
        la terminal (FR-012).
      parameters:
        - { name: id_sucursal, in: query, required: true, schema: { type: integer } }
        - name: solo_con_senal
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: Si es `true`, solo las terminales desactualizadas, expuestas o con versión de referencia desconocida.
      responses:
        '200':
          description: Terminales con su estado de firmware
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TerminalPago' }
        '400':
          description: '`pagos_sucursal_requerida`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /pagos/terminales/{id_terminal_pago}:
    patch:
      tags: [terminales]
      summary: Mover una terminal a otra sucursal, o retirarla
      description: |
        FR-014, FR-015. Mover cierra el tramo vigente de `historial_ubicacion` y abre uno nuevo; una
        señal de firmware se atribuye a la sucursal del tramo vigente EN EL MOMENTO EVALUADO. Retirar
        marca `activa = false` — NUNCA se borra la terminal (Principio IV). Anexa `terminal_movida` a
        la bitácora.
      parameters:
        - { name: id_terminal_pago, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_operador]
              properties:
                id_sucursal_destino: { type: integer, nullable: true, description: Presente para un movimiento. }
                fecha_movimiento: { type: string, format: date, nullable: true }
                retirar: { type: boolean, default: false }
                id_operador: { type: integer }
      responses:
        '200':
          description: Terminal actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TerminalPago' }
        '404':
          description: '`pagos_terminal_no_existe`, `pagos_sucursal_no_existe`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  schemas:
...
    Error:
      type: object
      required: [codigo, mensaje]
      properties:
        codigo: { type: string }
        mensaje:
          type: string
          description: Redactado para el operador, con la acción correctiva. Nunca una traza técnica.
...
    TerminalPago:
      type: object
      required:
        [id_terminal_pago, identificador, modelo, id_sucursal, version_firmware, activa,
         desactualizada, expuesta_a_clonacion, version_referencia_desconocida]
      properties:
        id_terminal_pago: { type: integer }
        identificador: { type: string }
        modelo: { type: string }
        id_sucursal: { type: integer, description: Sucursal actual. }
        version_firmware: { type: string }
        fecha_ultima_actualizacion_firmware: { type: string, format: date, nullable: true }
        ultima_version_referencia:
          type: string
          nullable: true
          description: 'De `ULTIMA_VERSION_FIRMWARE[modelo]` (config). null ⇒ version_referencia_desconocida.'
        desactualizada:
          type: boolean
          description: 'version_firmware < ultima_version_referencia (FR-010).'
        expuesta_a_clonacion:
          type: boolean
          description: 'version o modelo en LISTA_FIRMWARE_VULNERABLE (FR-011).'
        version_referencia_desconocida:
          type: boolean
          description: 'No hay ultima_version_referencia para el modelo — NUNCA se asume "al día" (FR-010).'
        referencias_vulnerabilidad:
          type: array
          items: { type: string }
          description: 'Referencias (CVE, boletín, nota interna) que motivan expuesta_a_clonacion. Vacío si no hay.'
        historial_ubicacion:
          type: array
          items:
            type: object
            properties:
              id_sucursal: { type: integer }
              desde: { type: string, format: date }
              hasta: { type: string, format: date, nullable: true }
        historial_firmware:
          type: array
          items:
            type: object
            properties:
              version: { type: string }
              fecha: { type: string, format: date }
              id_operador: { type: integer }
        activa: { type: boolean }
...